refactor(checker): split DeferredFlowEnvWrite out of def_mapping to hold the 2000-line cap
Adding the value-only-lib-shadow fix pushed def_mapping.rs over the 2000
physical-line arch-guard cap. Move the self-contained `DeferredFlowEnvWrite`
enum, its `apply` replay method, and `apply_augmented_def` into a sibling
`context/deferred_flow_env_write` module (a distinct concern: deferred
flow-environment write replay), and re-export `DeferredFlowEnvWrite` from
`context`. Behavior-preserving; no semantic change.
