+++ crates/tsz-checker/src/context/mod.rs
@@ -34,7 +34,8 @@ mod file_session_reset;
 pub mod lifetime_shells;
 pub use lifetime_shells::{FileSession, LspPersistentCache, SpeculationScope, WorkerContext};
 mod def_mapping;
-pub use def_mapping::DeferredFlowEnvWrite;
+mod deferred_flow_env_write;
+pub use deferred_flow_env_write::DeferredFlowEnvWrite;
 mod file_format_lookup;
 pub(crate) use file_format_lookup::{lookup_file_is_esm_in_map, lookup_is_external_module_in_map};
 mod import_conflicts;
